move vrf struct to its own file
